requirements.txt
Django
djangorestframework
data
django-filter
drf-yasg